// src/api/models/portfolioDetailResDTOHopeJob.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */

/**
 * @nullable
 */
export type PortfolioDetailResDTOHopeJob = typeof PortfolioDetailResDTOHopeJob[keyof typeof PortfolioDetailResDTOHopeJob] | null;


export const PortfolioDetailResDTOHopeJob = {
  NONE: 'NONE',
  PLANNING: 'PLANNING',
  MARKETING: 'MARKETING',
  DESIGN: 'DESIGN',
  DEV: 'DEV',
  MEDIA: 'MEDIA',
  DATA: 'DATA',
} as const;